+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - gitignore
  - t_tree_node [rbtree]
  - ft_tree_node_new [rbtree]
+ - ft_tree_node_insert [rbtree]
 
 ### Changed
 ### Fixed